% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crossmatch_example}
\alias{crossmatch_example}
\title{Example crossmatch dataset}
\format{
A tibble with 1,354 rows and 4 columns:
\describe{
\item{patient_id}{Unique identifier for each waitlisted patient}
\item{donor_id}{Unique identifier for each donor}
\item{hla_mismatch}{HLA mismatch score (ABDRDQ score) for this patient-donor pair. Calculated as: sqrt(1 x A_mismatches + 1.5 x B_mismatches + 3 x DR_mismtaches + 3 x DQ_mismatches)}
\item{unacceptable_antigens}{Presence of any donor HLA antigens that would be unnaceptable for this patient. 0 or 1}
}
}
\source{
Simulated data
}
\usage{
crossmatch_example
}
\description{
Example crossmatch dataset
}
\keyword{datasets}
