% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{donor_example}
\alias{donor_example}
\title{Example donor dataset}
\format{
A tibble with 1 rows and 14 columns:
\describe{
\item{donor_id}{Unique identifier for each donor}
\item{donor_state}{Australian state where the donor was recovered. Options are: NSW, VIC, QLD, SA, WA}
\item{donor_bloodgroup}{ABO blood group of the donor}
\item{donor_age}{Donor age in years}
\item{donor_hypertension}{Donor history of hypertension. 0 or 1}
\item{donor_diabetes}{Donor history of diabetes. 0 or 1}
\item{donor_creatinine}{Terminal serum creatinine level}
\item{donor_strokedeath}{Death due to stroke. 0 or 1}
\item{donor_height}{Donor height in cm}
\item{donor_weight}{Donor weight in kg}
\item{donor_dcd}{Donation after circulatory death (DCD). 0 or 1}
\item{donor_kdpi}{Kidney Donor Profile Index (KDPI). Integer 1-100}
\item{donor_kidneys}{Number of kidneys available for allocation. 1 or 2}
\item{donor_pancreas}{Binary indicator for pancreas availability. 0 or 1}
\item{donor_living}{Living donor (altruistic or end of kidney exchange chain. 0 or 1)}
}
}
\source{
Simulated data
}
\usage{
donor_example
}
\description{
Example donor dataset
}
\keyword{datasets}
