% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_state_net_debts.R
\name{get_state_net_debts}
\alias{get_state_net_debts}
\title{Get state net debts}
\usage{
get_state_net_debts(state_debts)
}
\arguments{
\item{state_debts}{A dataframe/tibble with 5 rows (one for each state) and 2 columns:
\describe{
\item{state}{OrganMatch lab state: NSW, VIC, QLD, SA, or WA}
\item{debt}{Number of kidneys that state owes to the national pool. A positive debt means that state owes kidneys, a negative means that state is owed kidneys.}
}}
}
\value{
A dataframe/tibble with 20 rows, one for each state paired with each other state (including itself), and 3 columns:
\describe{
\item{from_state}{The state kidneys are coming from: NSW, VIC, QLD, SA, or WA}
\item{to_state}{The state kidneys are going to: NSW, VIC, QLD, SA, or WA}
\item{net_debt}{The net number of kidneys the from_state owes to the to_state. Positive means the from_state owes kidneys, negative means the from_state is owed kidneys. If from_state is same as to_state, then net_debt is always 0.}
}
}
\description{
Function to calculate net kidney debts between each pair of OrganMatch lab states, from a given set of state kidney debts.
}
\examples{
state_debts <- sim_state_debts(min_debt = -6, max_debt = 6) # Simulate state debts from a uniform distribution, ensuring sum of debts across all states is zero.
state_net_debts <- get_state_net_debts(state_debts)
state_net_debts

}
