% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_debts.R
\name{sim_state_debts}
\alias{sim_state_debts}
\title{Simulate state kidney debts}
\usage{
sim_state_debts(min_debt, max_debt)
}
\arguments{
\item{min_debt}{Minimum possible debt (should be negative to allow at least some states to be owed kidneys)}

\item{max_debt}{Maximum possible debt (should be positive to allow at least some states to owe kidneys)}
}
\value{
A tibble with 5 rows (one for each state) and 2 columns:
\describe{
\item{state}{OrganMatch lab state: NSW, VIC, QLD, SA, or WA}
\item{debt}{Simulated net kidney debt for that state.}
}
}
\description{
Function to simulate number of kidneys owing/owed to each OrganMatch lab state (NSW, VIC, QLD, SA, WA). Samples from a uniform distribution, but ensures that sum of debts across all states is zero.
}
\examples{
# Assume states can be owed at most 6 kidneys (min_debt = -6), and can owe at most 6 kidneys (max_debt = 6)
state_debts <- sim_state_debts(min_debt = -6, max_debt = 6)

}
