% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_debts_example}
\alias{state_debts_example}
\title{Example state_debts dataset}
\format{
A tibble with 5 rows and 2 columns:
\describe{
\item{state}{State. Options are NSW, VIC, QLD, SA, WA}
\item{debt}{Net number of kidneys owed. Integer, can be negative.}
}
}
\source{
Simulated data
}
\usage{
state_debts_example
}
\description{
Example state_debts dataset
}
\keyword{datasets}
