% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{waitlist_example}
\alias{waitlist_example}
\title{Example waitlist dataset}
\format{
A tibble with 1,354 rows and 14 columns:
\describe{
\item{patient_id}{Unique identifier for each waitlisted patient}
\item{patient_waityears}{Total time on the waiting list in years (as a decimal, not rounded)}
\item{patient_state}{Australian state where the patient is waitlisted. Options are: NSW, VIC, QLD, SA, WA}
\item{patient_age}{Patient age in years}
\item{patient_bloodgroup}{ABO blood group of the patient}
\item{patient_pra}{Panel reactive antibody (PRA) value. Decimal 0-100}
\item{patient_epts}{Estimated Post-Transplant Survival (EPTS) score. Integer 1-100}
\item{patient_hla_mismatch_mean}{Mean HLA mismatch score}
\item{patient_hla_mismatch_sd}{Standard deviation of HLA mismatch score}
\item{patient_national_urgent}{Indicator for nationally urgent status. 0 or 1}
\item{patient_state_urgent}{Indicator for state urgent status. 0 or 1}
\item{patient_prior_donor}{Indicator for prior living donor status. 0 or 1}
\item{patient_kidney_after_other_organ}{Indicator for kidney-after-other-organ transplant. 0 or 1}
\item{patient_spk}{Indicator for simultaneous pancreas–kidney transplant. 0 or 1}
}
}
\source{
Simulated data
}
\usage{
waitlist_example
}
\description{
Example waitlist dataset
}
\keyword{datasets}
