% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_debts_example}
\alias{state_debts_example}
\title{Example state_debts dataset}
\format{
A tibble with 25 rows and 3 columns:
\describe{
\item{from_state}{State that owes kidneys. Option are NSW, VIC, QLD, SA, WA}
\item{to_state}{State that is owed kidneys. Option are NSW, VIC, QLD, SA, WA}
\item{net_debt}{Net number of kidneys owed. Integer, can be negative. Must by symmetric, e.g. if NSW -> VIC net debt is +2, then VIC -> NSW net debt must be -2. if from_state = to_state then net_debt must be 0.}
}
}
\source{
Simulated data
}
\usage{
state_debts_example
}
\description{
Example state_debts dataset
}
\keyword{datasets}
